Move subshell package to pkg dir
