+++ pkg/subshell/bash/bash_test.go
