+++ pkg/subshell/cmd/cmd_test.go
