+++ pkg/subshell/fish/fish.go
@@ -1,35 +1,26 @@
-package bash
+package fish
 
 import (
+	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"runtime"
 
 	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/internal/errs"
 	"github.com/ActiveState/cli/internal/fileutils"
-	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/logging"
 	"github.com/ActiveState/cli/internal/osutils"
 	"github.com/ActiveState/cli/internal/osutils/user"
 	"github.com/ActiveState/cli/internal/output"
-	"github.com/ActiveState/cli/internal/subshell/sscommon"
 	"github.com/ActiveState/cli/pkg/project"
+	sscommon2 "github.com/ActiveState/cli/pkg/subshell/sscommon"
 )
 
 var escaper *osutils.ShellEscape
 
-var rcFileName = ".bashrc"
-
 func init() {
 	escaper = osutils.NewBashEscaper()
-
-	// On macOS all terminal windows run login shells, this means that
-	// .bashrc can be ignored so we instead use .bash_profile
-	if runtime.GOOS == "darwin" {
-		rcFileName = ".bash_profile"
-	}
 }
 
 // SubShell covers the subshell.SubShell interface, reference that for documentation
@@ -41,7 +32,7 @@ type SubShell struct {
 	errs   chan error
 }
 
-const Name string = "bash"
+const Name string = "fish"
 
 // Shell - see subshell.SubShell
 func (v *SubShell) Shell() string {
@@ -59,57 +50,48 @@ func (v *SubShell) SetBinary(binary string) {
 }
 
 // WriteUserEnv - see subshell.SubShell
-func (v *SubShell) WriteUserEnv(cfg sscommon.Configurable, env map[string]string, envType sscommon.RcIdentification, _ bool) error {
+func (v *SubShell) WriteUserEnv(cfg sscommon2.Configurable, env map[string]string, envType sscommon2.RcIdentification, _ bool) error {
 	rcFile, err := v.RcFile()
 	if err != nil {
 		return errs.Wrap(err, "RcFile failure")
 	}
 
-	if path, pathExists := env["PATH"]; pathExists && runtime.GOOS == "windows" {
-		bashified, err := osutils.BashifyPathEnv(path)
-		if err != nil {
-			return errs.Wrap(err, "Unable to bashify PATH: %v", path)
-		}
-		env["PATH"] = bashified
-	}
-
-	env = sscommon.EscapeEnv(env)
-	return sscommon.WriteRcFile("bashrc_append.sh", rcFile, envType, env)
+	env = sscommon2.EscapeEnv(env)
+	return sscommon2.WriteRcFile("fishrc_append.fish", rcFile, envType, env)
 }
 
-func (v *SubShell) CleanUserEnv(cfg sscommon.Configurable, envType sscommon.RcIdentification, _ bool) error {
+func (v *SubShell) CleanUserEnv(cfg sscommon2.Configurable, envType sscommon2.RcIdentification, _ bool) error {
 	rcFile, err := v.RcFile()
 	if err != nil {
-		return errs.Wrap(err, "RcFile-failure")
+		return errs.Wrap(err, "RcFile failure")
 	}
 
-	if err := sscommon.CleanRcFile(rcFile, envType); err != nil {
+	if err := sscommon2.CleanRcFile(rcFile, envType); err != nil {
 		return errs.Wrap(err, "Failed to remove %s from rcFile", envType)
 	}
 
 	return nil
 }
 
-func (v *SubShell) RemoveLegacyInstallPath(cfg sscommon.Configurable) error {
+func (v *SubShell) RemoveLegacyInstallPath(cfg sscommon2.Configurable) error {
 	rcFile, err := v.RcFile()
 	if err != nil {
 		return errs.Wrap(err, "RcFile-failure")
 	}
 
-	return sscommon.RemoveLegacyInstallPath(rcFile)
+	return sscommon2.RemoveLegacyInstallPath(rcFile)
 }
 
 func (v *SubShell) WriteCompletionScript(completionScript string) error {
-	dir := "/usr/local/etc/bash_completion.d/"
-	if runtime.GOOS != "darwin" {
-		dir = "/etc/bash_completion.d/"
+	homeDir, err := user.HomeDir()
+	if err != nil {
+		return errs.Wrap(err, "IO failure")
 	}
 
-	fpath := filepath.Join(dir, constants.CommandName)
-	logging.Debug("Writing to %s: %s", fpath, completionScript)
-	err := fileutils.WriteFile(fpath, []byte(completionScript))
+	fpath := filepath.Join(homeDir, ".config/fish/completions", constants.CommandName+".fish")
+	err = fileutils.WriteFile(fpath, []byte(completionScript))
 	if err != nil {
-		logging.Debug("Could not write completions script '%s', likely due to non-admin privileges", fpath)
+		return errs.Wrap(err, "Could not write completions script")
 	}
 
 	return nil
@@ -121,7 +103,7 @@ func (v *SubShell) RcFile() (string, error) {
 		return "", errs.Wrap(err, "IO failure")
 	}
 
-	return filepath.Join(homeDir, rcFileName), nil
+	return filepath.Join(homeDir, ".config/fish/config.fish"), nil
 }
 
 func (v *SubShell) EnsureRcFileExists() error {
@@ -135,20 +117,12 @@ func (v *SubShell) EnsureRcFileExists() error {
 
 // SetupShellRcFile - subshell.SubShell
 func (v *SubShell) SetupShellRcFile(targetDir string, env map[string]string, namespace *project.Namespaced) error {
-	env = sscommon.EscapeEnv(env)
-	return sscommon.SetupShellRcFile(filepath.Join(targetDir, "shell.sh"), "bashrc_global.sh", env, namespace)
+	env = sscommon2.EscapeEnv(env)
+	return sscommon2.SetupShellRcFile(filepath.Join(targetDir, "shell.fish"), "fishrc_global.fish", env, namespace)
 }
 
 // SetEnv - see subshell.SetEnv
 func (v *SubShell) SetEnv(env map[string]string) error {
-	if path, pathExists := env["PATH"]; pathExists && runtime.GOOS == "windows" {
-		bashified, err := osutils.BashifyPathEnv(path)
-		if err != nil {
-			return locale.WrapError(err, "err_unable_set_bashify_PATH", "Unable to setup bash-style PATH")
-		}
-		env["PATH"] = bashified
-	}
-
 	v.env = env
 	return nil
 }
@@ -159,25 +133,27 @@ func (v *SubShell) Quote(value string) string {
 }
 
 // Activate - see subshell.SubShell
-func (v *SubShell) Activate(proj *project.Project, cfg sscommon.Configurable, out output.Outputer) error {
+func (v *SubShell) Activate(proj *project.Project, cfg sscommon2.Configurable, out output.Outputer) error {
 	var shellArgs []string
 	var directEnv []string
 
 	// available project files require more intensive modification of shell envs
 	if proj != nil {
-		env := sscommon.EscapeEnv(v.env)
+		env := sscommon2.EscapeEnv(v.env)
 		var err error
-		if v.rcFile, err = sscommon.SetupProjectRcFile(proj, "bashrc.sh", "", env, out, cfg, true); err != nil {
+		if v.rcFile, err = sscommon2.SetupProjectRcFile(proj, "fishrc.fish", "", env, out, cfg, false); err != nil {
 			return err
 		}
 
-		shellArgs = append(shellArgs, "--rcfile", v.rcFile.Name())
+		shellArgs = append(shellArgs,
+			"-i", "-C", fmt.Sprintf("source %s", v.rcFile.Name()),
+		)
 	} else {
-		directEnv = sscommon.EnvSlice(v.env)
+		directEnv = sscommon2.EnvSlice(v.env)
 	}
 
-	cmd := sscommon.NewCommand(v.Binary(), shellArgs, directEnv)
-	v.errs = sscommon.Start(cmd)
+	cmd := sscommon2.NewCommand(v.Binary(), shellArgs, directEnv)
+	v.errs = sscommon2.Start(cmd)
 	v.cmd = cmd
 	return nil
 }
@@ -193,7 +169,7 @@ func (v *SubShell) Deactivate() error {
 		return nil
 	}
 
-	if err := sscommon.Stop(v.cmd); err != nil {
+	if err := sscommon2.Stop(v.cmd); err != nil {
 		return err
 	}
 
@@ -203,7 +179,7 @@ func (v *SubShell) Deactivate() error {
 
 // Run - see subshell.SubShell
 func (v *SubShell) Run(filename string, args ...string) error {
-	return sscommon.RunFuncByBinary(v.Binary())(osutils.EnvMapToSlice(v.env), filename, args...)
+	return sscommon2.RunFuncByBinary(v.Binary())(osutils.EnvMapToSlice(v.env), filename, args...)
 }
 
 // IsActive - see subshell.SubShell
