+++ pkg/subshell/fish/fish_test.go
