+++ pkg/subshell/sscommon/rcfile_test.go
