+++ pkg/subshell/sscommon/sscommon_darwin.go
