+++ pkg/subshell/sscommon/sscommon_lin_mac.go
