+++ pkg/subshell/zsh/zsh.go
@@ -1,20 +1,24 @@
-package fish
+package zsh
 
 import (
 	"fmt"
+	"io/ioutil"
+	"log"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strings"
 
 	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/internal/errs"
 	"github.com/ActiveState/cli/internal/fileutils"
+	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/logging"
 	"github.com/ActiveState/cli/internal/osutils"
 	"github.com/ActiveState/cli/internal/osutils/user"
 	"github.com/ActiveState/cli/internal/output"
-	"github.com/ActiveState/cli/internal/subshell/sscommon"
 	"github.com/ActiveState/cli/pkg/project"
+	sscommon2 "github.com/ActiveState/cli/pkg/subshell/sscommon"
 )
 
 var escaper *osutils.ShellEscape
@@ -32,7 +36,7 @@ type SubShell struct {
 	errs   chan error
 }
 
-const Name string = "fish"
+const Name string = "zsh"
 
 // Shell - see subshell.SubShell
 func (v *SubShell) Shell() string {
@@ -50,48 +54,63 @@ func (v *SubShell) SetBinary(binary string) {
 }
 
 // WriteUserEnv - see subshell.SubShell
-func (v *SubShell) WriteUserEnv(cfg sscommon.Configurable, env map[string]string, envType sscommon.RcIdentification, _ bool) error {
+func (v *SubShell) WriteUserEnv(cfg sscommon2.Configurable, env map[string]string, envType sscommon2.RcIdentification, _ bool) error {
 	rcFile, err := v.RcFile()
 	if err != nil {
 		return errs.Wrap(err, "RcFile failure")
 	}
 
-	env = sscommon.EscapeEnv(env)
-	return sscommon.WriteRcFile("fishrc_append.fish", rcFile, envType, env)
+	env = sscommon2.EscapeEnv(env)
+	return sscommon2.WriteRcFile("zshrc_append.sh", rcFile, envType, env)
 }
 
-func (v *SubShell) CleanUserEnv(cfg sscommon.Configurable, envType sscommon.RcIdentification, _ bool) error {
+func (v *SubShell) CleanUserEnv(cfg sscommon2.Configurable, envType sscommon2.RcIdentification, _ bool) error {
 	rcFile, err := v.RcFile()
 	if err != nil {
 		return errs.Wrap(err, "RcFile failure")
 	}
 
-	if err := sscommon.CleanRcFile(rcFile, envType); err != nil {
+	if err := sscommon2.CleanRcFile(rcFile, envType); err != nil {
 		return errs.Wrap(err, "Failed to remove %s from rcFile", envType)
 	}
 
 	return nil
 }
 
-func (v *SubShell) RemoveLegacyInstallPath(cfg sscommon.Configurable) error {
+func (v *SubShell) RemoveLegacyInstallPath(cfg sscommon2.Configurable) error {
 	rcFile, err := v.RcFile()
 	if err != nil {
 		return errs.Wrap(err, "RcFile-failure")
 	}
 
-	return sscommon.RemoveLegacyInstallPath(rcFile)
+	return sscommon2.RemoveLegacyInstallPath(rcFile)
 }
 
 func (v *SubShell) WriteCompletionScript(completionScript string) error {
+	dir := "/usr/local/share/zsh/site-functions"
+	if fpath := os.Getenv("FPATH"); fpath != "" {
+		fpathv := strings.Split(fpath, ":")
+		if len(fpathv) > 0 {
+			dir = fpathv[0]
+		}
+	}
+
+	fpath := filepath.Join(dir, "_"+constants.CommandName)
+	logging.Debug("Writing to %s: %s", fpath, completionScript)
+	err := fileutils.WriteFile(fpath, []byte(completionScript))
+	if err != nil {
+		return errs.Wrap(err, "Could not write completions script")
+	}
+
 	homeDir, err := user.HomeDir()
 	if err != nil {
 		return errs.Wrap(err, "IO failure")
 	}
 
-	fpath := filepath.Join(homeDir, ".config/fish/completions", constants.CommandName+".fish")
-	err = fileutils.WriteFile(fpath, []byte(completionScript))
-	if err != nil {
-		return errs.Wrap(err, "Could not write completions script")
+	// Remove the zsh completions cache so our completion script actually gets picked up
+	if err := os.Remove(filepath.Join(homeDir, ".zcompdump")); err != nil {
+		// non-critical, we're just trying to eliminate any issues caused by zsh's caching
+		logging.Debug("Could not delete .zcompdump: %v", err)
 	}
 
 	return nil
@@ -103,7 +122,7 @@ func (v *SubShell) RcFile() (string, error) {
 		return "", errs.Wrap(err, "IO failure")
 	}
 
-	return filepath.Join(homeDir, ".config/fish/config.fish"), nil
+	return filepath.Join(homeDir, ".zshrc"), nil
 }
 
 func (v *SubShell) EnsureRcFileExists() error {
@@ -117,8 +136,8 @@ func (v *SubShell) EnsureRcFileExists() error {
 
 // SetupShellRcFile - subshell.SubShell
 func (v *SubShell) SetupShellRcFile(targetDir string, env map[string]string, namespace *project.Namespaced) error {
-	env = sscommon.EscapeEnv(env)
-	return sscommon.SetupShellRcFile(filepath.Join(targetDir, "shell.fish"), "fishrc_global.fish", env, namespace)
+	env = sscommon2.EscapeEnv(env)
+	return sscommon2.SetupShellRcFile(filepath.Join(targetDir, "shell.zsh"), "zshrc_global.sh", env, namespace)
 }
 
 // SetEnv - see subshell.SetEnv
@@ -133,27 +152,52 @@ func (v *SubShell) Quote(value string) string {
 }
 
 // Activate - see subshell.SubShell
-func (v *SubShell) Activate(proj *project.Project, cfg sscommon.Configurable, out output.Outputer) error {
-	var shellArgs []string
+func (v *SubShell) Activate(proj *project.Project, cfg sscommon2.Configurable, out output.Outputer) error {
 	var directEnv []string
 
 	// available project files require more intensive modification of shell envs
 	if proj != nil {
-		env := sscommon.EscapeEnv(v.env)
+		env := sscommon2.EscapeEnv(v.env)
 		var err error
-		if v.rcFile, err = sscommon.SetupProjectRcFile(proj, "fishrc.fish", "", env, out, cfg, false); err != nil {
+		if v.rcFile, err = sscommon2.SetupProjectRcFile(proj, "zshrc.sh", "", env, out, cfg, false); err != nil {
 			return err
 		}
 
-		shellArgs = append(shellArgs,
-			"-i", "-C", fmt.Sprintf("source %s", v.rcFile.Name()),
-		)
+		path, err := ioutil.TempDir("", "state-zsh")
+		if err != nil {
+			return errs.Wrap(err, "OS failure")
+		}
+
+		activeZsrcPath := filepath.Join(path, ".zshrc")
+		err = fileutils.CopyFile(v.rcFile.Name(), activeZsrcPath)
+		if err != nil {
+			return err
+		}
+
+		// If users have set $ZDOTDIR then we need to make sure their zshrc file uses it
+		// and if it hasn't been set, user $HOME as that is often a default for zsh setup
+		// commands.
+		userzdotdir := os.Getenv("ZDOTDIR")
+		if userzdotdir == "" {
+			homeDir, err := user.HomeDir()
+			if err != nil {
+				log.Println(locale.T("Warning: Could not load home directory for current user"))
+			} else {
+				userzdotdir = homeDir
+			}
+		}
+
+		err = fileutils.PrependToFile(activeZsrcPath, []byte(fmt.Sprintf("export ZDOTDIR=%s\n", userzdotdir)))
+		if err != nil {
+			return err
+		}
+		os.Setenv("ZDOTDIR", path)
 	} else {
-		directEnv = sscommon.EnvSlice(v.env)
+		directEnv = sscommon2.EnvSlice(v.env)
 	}
 
-	cmd := sscommon.NewCommand(v.Binary(), shellArgs, directEnv)
-	v.errs = sscommon.Start(cmd)
+	cmd := sscommon2.NewCommand(v.Binary(), nil, directEnv)
+	v.errs = sscommon2.Start(cmd)
 	v.cmd = cmd
 	return nil
 }
@@ -169,7 +213,7 @@ func (v *SubShell) Deactivate() error {
 		return nil
 	}
 
-	if err := sscommon.Stop(v.cmd); err != nil {
+	if err := sscommon2.Stop(v.cmd); err != nil {
 		return err
 	}
 
@@ -179,7 +223,7 @@ func (v *SubShell) Deactivate() error {
 
 // Run - see subshell.SubShell
 func (v *SubShell) Run(filename string, args ...string) error {
-	return sscommon.RunFuncByBinary(v.Binary())(osutils.EnvMapToSlice(v.env), filename, args...)
+	return sscommon2.RunFuncByBinary(v.Binary())(osutils.EnvMapToSlice(v.env), filename, args...)
 }
 
 // IsActive - see subshell.SubShell
