+++ pkg/subshell/zsh/zsh_test.go
